+++ CHANGELOG.md
@@ -182,6 +182,7 @@ This release brings lots of Sheet changes to fix a lot of wonk exposed by the fo
 - Migrating actors that are on scenes in a compendium would fail because the code did UUID lookups and Foundry won't let you do this synchronously. Added code so if the restoration is for a Compenium actor/item, the lookup is done with an async await.
 - Fixed an issue where if migration performs a backup/restore of an installed item, it would delete the original item which was causing the container item to report the incorrect amount of used slots.
 - Standardized the passed migration context to actor:*EmbeddedDocuments to isMigrating
+- #705 - When clicking the DV button on a sheet, users would sometimes feel like the Ruler was broken if they did not have the token selected. A warning message is now thrown if a user clicks the DV button and they do not have a token selected.  Additionally, the DV ruler is now highlighted for the ruler that is currently active.
 
 ## Version 0.86.1 | Date: 2023-02-05